+++ simulator/src/components/Simulator.js
@@ -502,7 +502,7 @@ module.exports =   React.createClass({
                 options = {};
             }
             options.reLinkWidgets = true;
-            this.handleTargetAction(page, 'Load');
+            
 
             switch (method){
                 case 'MOVE_LR':
@@ -520,7 +520,8 @@ module.exports =   React.createClass({
                     }.bind(this),function () {
                         page.translate = null;
                         options.pageAnimate = false
-                    })
+                        this.handleTargetAction(page, 'Load');
+                    }.bind(this))
                     break;
                 case 'MOVE_RL':
                     AnimationManager.step(offcanvas.width,0,0,0,duration,frames,easing,function (deltas) {
@@ -537,7 +538,8 @@ module.exports =   React.createClass({
                     }.bind(this),function () {
                         page.translate = null;
                         options.pageAnimate = false
-                    })
+                        this.handleTargetAction(page, 'Load');
+                    }.bind(this))
                     break;
                 case 'MOVE_TB':
                     AnimationManager.step(-offcanvas.height,0,0,0,duration,frames,easing,function (deltas) {
@@ -554,7 +556,8 @@ module.exports =   React.createClass({
                     }.bind(this),function () {
                         page.translate = null;
                         options.pageAnimate = false
-                    })
+                        this.handleTargetAction(page, 'Load');
+                    }.bind(this))
                     break;
                 case 'MOVE_BT':
                     AnimationManager.step(offcanvas.height,0,0,0,duration,frames,easing,function (deltas) {
@@ -571,7 +574,8 @@ module.exports =   React.createClass({
                     }.bind(this),function () {
                         page.translate = null;
                         options.pageAnimate = false;
-                    })
+                        this.handleTargetAction(page, 'Load');
+                    }.bind(this))
                     break;
                 case 'SCALE':
                     var beforeTranslateMatrix = [
@@ -609,12 +613,14 @@ module.exports =   React.createClass({
                     }.bind(this),function () {
                         page.transform = null
                         options.pageAnimate = false;
-                    })
+                        this.handleTargetAction(page, 'Load');
+                    }.bind(this))
 
 
 
                     break;
                 default:
+                    this.handleTargetAction(page, 'Load');
                     this.draw(null,options);
             }
 
@@ -1154,7 +1160,7 @@ module.exports =   React.createClass({
             // subCanvas.info.w = w;
             // subCanvas.info.h = h;
             //generate load trigger
-            this.handleTargetAction(subCanvas, 'Load')
+            
             //transition animation
             var moveX = w;
             var moveY = 0;
@@ -1181,7 +1187,8 @@ module.exports =   React.createClass({
                         }.bind(this),function () {
                             // offctx.restore()
                             subCanvas.translate = null;
-                        })
+                            this.handleTargetAction(subCanvas, 'Load')
+                        }.bind(this))
                         break;
                     case 'MOVE_RL':
                         AnimationManager.step(w,0,0,0,duration,frames,easing,function (deltas) {
@@ -1198,7 +1205,8 @@ module.exports =   React.createClass({
                         }.bind(this),function () {
                             // offctx.restore()
                             subCanvas.translate = null;
-                        })
+                            this.handleTargetAction(subCanvas, 'Load')
+                        }.bind(this))
                         break;
                     case 'SCALE':
                         var beforeTranslateMatrix = [
@@ -1234,12 +1242,14 @@ module.exports =   React.createClass({
                             this.draw(null,options);
                         }.bind(this),function () {
                             subCanvas.transform = null
-                        })
+                            this.handleTargetAction(subCanvas, 'Load')
+                        }.bind(this))
 
 
 
                         break;
                     default:
+                        this.handleTargetAction(subCanvas, 'Load')
                         this.drawSingleSubCanvas(subCanvas, x, y, w, h, options)
 
                 }