+++ changelog.md
@@ -130,4 +130,4 @@ fix:
     2. mulitiColor progress's cursor not display in simulator
     3. remove Slide and Switch widget's action
 add:
-    1. add change subLayer zIndex fun
+    1. add change subLayer zIndex fun