+++ changelog.md
@@ -129,5 +129,6 @@ fix:
     1. copy page bug 
     2. mulitiColor progress's cursor not display in simulator
     3. remove Slide and Switch widget's action
+    4. change some anmiation logic
 add:
     1. add change subLayer zIndex fun